# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shell-Tunnel API",
    "description": "Ultra-lightweight remote shell gateway. Provides a cross-platform REST/WebSocket API for programmatic interaction with system shells.",
    "version": "0.7.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "iyulab",
      "url": "https://github.com/iyulab/shell-tunnel"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/sessions/{sessionId}/execute": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute command in session",
        "description": "Executes a command in an existing session.\n\nRequired capability: `exec`.",
        "operationId": "executeInSession",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          }
        ],
        "requestBody": {
          "description": "Command to execute",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/execute": {
      "post": {
        "tags": [
          "Execution"
        ],
        "summary": "Execute command (one-shot)",
        "description": "Executes a command without creating a persistent session.\n\nRequired capability: `exec`.",
        "operationId": "executeOneshot",
        "requestBody": {
          "description": "Command to execute",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque bearer token via the Authorization header. Each token carries a set of capabilities (exec, session.read, session.manage; '*' wildcard). A route declares a required capability: a missing/invalid token yields 401, a valid token lacking the capability yields 403. Role presets: operator, read-only, full-control."
      }
    },
    "parameters": {
      "sessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "description": "Session identifier",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "ExecuteCommandRequest": {
        "type": "object",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "Command to execute",
            "example": "ls -la"
          },
          "working_dir": {
            "type": "string",
            "description": "Working directory override"
          },
          "timeout_secs": {
            "type": "integer",
            "description": "Command timeout in seconds",
            "default": 30,
            "minimum": 1,
            "maximum": 300
          }
        }
      },
      "ExecuteCommandResponse": {
        "type": "object",
        "required": [
          "output",
          "exit_code",
          "duration_ms"
        ],
        "properties": {
          "output": {
            "type": "string",
            "description": "Command output (stdout + stderr)"
          },
          "exit_code": {
            "type": "integer",
            "description": "Exit code (0 for success)"
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Execution duration in milliseconds"
          },
          "timed_out": {
            "type": "boolean",
            "description": "Whether the command timed out",
            "default": false
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Command cannot be empty"
          },
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": true
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required: the Authorization: Bearer token is missing or invalid (empty body)."
      },
      "ForbiddenError": {
        "description": "The bearer token is valid but lacks the capability this route requires (empty body)."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}